Renamed "Credentials" directory to "credentials"
