+++ packages/realm-web/src/credentials/AnonymousCredentials.ts
